requirements.txt
aiogram>=3.4
aiosqlite
python-dotenv
apscheduler>=3.10
